[OM-61086] Added support to enable/disable telemetry based on user input
Now the cron job can use the response it gets from group component
regarding whether the user has enabled or not telemetry from the UI and
update Prometheus configuration as needed to start/stop sending data to
DataCloud. Metrics will still be scraped by Prometheus: they will just
not be sent to the DataCloud gateway container and by extension, not to
DataCloud.

The 'safe' variants of `load` and `dump` were used for manipulating YAML
to avoid potential security risks.

Built and deployed prometheus-config-maganger to an XL VM in our lab and
verified that:

- When I enable telemetry from the UI, the DataCloud remote writer
  configuration is present in the Prometheus configuration file.
  When I disable it, it's not.
- When I enable telemetry from the UI, I see log lines from DataCloud
  gateway container about sending data. When I disable it, I don't.
- The cron jobs complete successfully and without errors in the logs.

Also:

- Added comments and TODOs to the Python script.
- Removed code intended solely for sanity checks during initial
  development.
- Added some rudimentary logging.

The script still needs error handling code and a number of strings like
paths and URLs need to be converted to script/container arguments and
Helm values.