[OM-61741]: Only perform cloud scaling capacity resizes on cloud scaling entities RB:41151
**Problem**
When converting commodities bought in the TopologyConverter, there are two code paths: one for on-prem entities, and one for cloud entities. The cloud entity code path adjusts the used value on commodities bought to be based on its "driving" commodity sold. For the cloud code path, we save information in the CommoditiesResizeTracker to determine whether a commodity was underutilized and use the CommoditiesResizeTracker information when generating explanations for actions on these cloud entities.

This codepath for converting cloud entities' commodities bought makes sense for entities for which we generate cloud scaling actions (where we select a new template for the entity to move to) such as VMs, Databases, and DatabaseServers. However, it is being incorrectly applied to ALL entities that report a cloud environment type, which includes Containers, ContainerPods, and all Application entities hosted by cloud entities.

There are two problems with this:
1. It incorrectly adjusts commodity bought used values for entities that should behave identically whether they are running in the cloud or not (Containers and Applications ideally are agnostic to whether the VM hosting them is in the cloud or on-prem).
2. Saving information in the CommoditiesResizeTracker results in incorrect explanations. The CommoditiesResizeTracker was written to support moves for VDI-related entities, but is applied generically to all entity types. This is fine if only cloud-scaling entities' information is stored in CommoditiesResizeTracker, but if ContainerPod information is in there, it results in bad explanations. This may also be affecting explanations for other entity types that the market generates Move actions for.

**Solution**
Only apply the logic for adjusting commodities bought to cloud-scaling entities, rather than applying it to all entities hosted in the cloud. I have renamed the "getOnPremResizeCapacity" method to "getStandardResizedCapacity" to reflect that it makes sense for entities hosted in the cloud as well as on-prem entities as long as they are not cloud-scaling entities. I reviewed this logic change with Thiru and Shravan who both advised this broader change rather than a more scoped change that used e.g. an EntityType blacklist to prevent cloud scaling logic adjusting commodities bought on Containers and ContainerPods hosted in the cloud.

Also adjusted a warning-level log message that was being very spammy to debug-level.

**Testing Done**
* Verified the correct actions are generated for cloud native entities and that the explanation for pod moves now reflect the correct reason for the move.
* Added a new unit test to exercise the new logic.