[OM-61384] Remove duplication within CloudCostStatsSubquery
RB:41053

The Storage Summary widget makes two costPrice stat requests: one grouped by storage tier, and one grouped by volume attachment status. For each of these stat requests, a query for CloudCostStatRecords is sent to the cost component. However, since group-by processing happens afterward, the query is identical for each stat request, as is the data returned. Since the raw data is identical for each stat request, only one query to the cost component needs to be made.

Also, iterating through CloudCostStatRecords is inefficient because the same tier-based subdivision and processing happens for each CloudCostStatRecord when it really only needs to happen for each timestamp within the CloudCostStatRecords. For this reason, combining the data from CloudCostStatRecords by associated timestamp reduces the number of iterations of this time-intensive task.