[OM-55465]: RB-40998: XL API: Fix 0 totalMin & totalMax values in market entities stats response
Background: Executing POST/markets/{market_Uuid}/entities/stats with a PM as scope and a
commodity that has byte/sec as unit (i.e. IOThroughput/NetThroughput), results in 0s for
totalMin and totalMax values. This is happenning because in the api component, during
the conversion to bit/sec that is needed for those commodities, totalMin & totalMax
values are being skipped.

Fix: Add totalMin & totalMax to the conversion.