[OM-61086] Added support for sending customer domain as telemetry label.
Modified prometheus configuration manager script to get the customer
domain from the license returned by the auth component and update
Prometheus configuration to include it in a label added to data sent to
DataCloud.

Also:

- Did a selective import to shorten some calls.
- Fixed a grammar mistake

Deployed image to an XL instance in our lab that had one valid and one
expired license and verified the string "turbonomic.com" appeared as the
value of the "customer_domain" label in DataCloud telemetry test project.