[OM-62020]: Only show merged WorkloadController actions on WorkloadControllers RB:41221
**Problem**
When the fix for OM_61839 was implemented (show ContainerSpec names in the WorkloadController action descriptions), a side effect was to have the WorkloadController actions start showing up on ContainerSpec entities. Ideally the merged WorkloadController should only show up on the WorkloadControllers. This will become even more confusing when we introduce merged actions on the ContainerSpecs themselves for the ProveIt UI.

**Solution**
The reason that these actions started showing up when scoped to ContainerSpecs is because in order to insert the name in the description, we had to list the ContainerSpecs as involved entities in the action. This is because the existing mechanism for including entity display names in actions is:
1. In the action orchestrator, include a format-string indicating that the API component should do string replacement for an entity when it does involved entity lookups. An example String would be like: "(^_^)~Underutilized VMem in Virtual Machine {entity:0:displayName:}". Then the API will look up the enitty with ID 0 and replace the String with the entity's displayName to become something like "Underutilized VMem in Virtual Machine Elasticsearch-0".
2. The API only looks up information for entities involved in the action. So if an explanation references an entity not involved in an action, the replacement cannot be performed and the name will be missing. However, a side effect of this is that the actions will be displayed in the UI on all involved entities. Thus, the behavior of mentioning an entity's display name in an explanation was tightly coupled to also showing that action when scoped to the other entity in the UI.

The reason the current implementation worked this way is because the API has to look up involved entities for an action for other purposes beyond just explanation generation. Thus, if the ActionOrchestrator also had to look up involved entity details, that would double the number of RPC calls we had to make to look up the information. However, Roman recently introduced a minimal topology graph in the ActionOrchestrator as an optimization which contains the information necessary for the explanations for realtime actions (plan action explanations won't have the minimal topology available). This review uses the minimal topology when possible to generate explanations, and as a fallback when the minimal topology doesn't have the explanation, we use the string-format substitution.

As a further optimization, we also use the minimal topology graph for action translation when converting MHz to VCPU counts for VCPU resizes on VMs. This avoids calls to the repository for Host CPU speed frequency which is already available in the minimal topology graph for realtime actions (plan will still have to call the repository).

**Testing Done**
* Observed the correct action explanations on workload controller actions.
* Workload Controller actions no longer show up on ContainerSpecs.
* Verified VM VCPU resize actions still show the correct action explanations and VCPU numbers.
* Saw performance of action translation improved (from about 3.2 seconds to 2.5 seconds in my local testing) and verified in Jaeger we are making fewer calls to the repository during action translation.
* Verified plan action descriptions are still correct as well. Since we do not generate merged WorkloadController actions in plans, there is no regression on these as a result of this fix.