[OM-62023]: XL AO: Ignore topology ID when retrieving entity settings for action processing.
Issue
=====
Action mode calculation always returning "MANUAL" at BofA for 7.22.6.

Background
==========
The topology processor uploads the per-entity settings breakdown to the group component at broadcast. The AO requests this breakdown when processing actions, and uses the settings for entities involved in each action to calculate the action mode. The AO request to the group component specified the topology ID: we wanted to get the entity settings for the same topology that the market ran the analysis on.

When BofA upgraded their prod instance to 7.22.6 the market slowed down from ~20mins to ~30 mins. The topology broadcast time stayed the same (a little under 10mins). This means there is now NO overlap between the topology processor's upload of the per-entity settings and the action orchestrator's retrieval of the per-entity settings for the same topology ID. The group component always returns an error, saying the per-entity settings are not found.

Fix
===
The fix is to remove the topology ID restriction when fetching settings from the AO. This means we will get the latest per-entity settings (even if they are on a slightly newer topology). This may have consequences if topology or settings change AND analysis is slow, so we are adding a flag to allow dynamic re-enabling of the strict topology ID matching.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/41222/