[OM-60296]: Add more validation for input placement policy
Intent
======
Add more validation for input placement policy

Problem
=======
The customer topology broadcast pipeline fails with an error which is caused by an invalid policy.

Diagnosis
=========
Looking at the logs showed that the customer created a policy with no type and that caused the issue.

Solution
========
Add more validation to the policy creation logic. This prevents from such errors to happen.