[OM-61946]: XL: Generate python protobuf and gRPC bindings in XL build.
Intent
======
Allow small python utilities to interface with other components' gRPC interfaces.

Background
==========
We've had multiple use cases for wanting to have a small lightweight python utility that can make a call to some components' gRPC server and parse the protobufs contained there. The thing I need it for is Grafana License Management. Nitya wanted to use it for the telemetry prometheus sidecar. It's also useful for various scripts.

Implementation
==============
The protobuf-maven-plugin we use to generate Java bindings does not support generating Python gRPC bindings out of the box. Instead of trying to figure out how to set that up, and further complicating the common-protobuf POM, I created a new module (https://git.turbonomic.com/turbonomic/python-grpc-generator) which will create a docker image, bundled with python, the grpc utilities, and a script to generate bindings.

The xl-python module will:
1) Use the maven dependency plugin to get the SDK and XL .proto files.
2) Use the fabric8 docker plugin to run the docker image from 1) against the files from 1).
3) Zip up the resulting python files and add them to the python-grpc artifact.

Reviewed By: https://rbcommons.com/s/VMTurbo/r/41153/