[OM-59892]: Restore all diags for group component in one transaction
Intent
======
Restore the diags for group component in one transaction

Problem
=======
Intermittently, PT is seeing an exception when they are reloading the diags with large customer topologies.

Diagnosis
=========
This is happening as broadcast happens while the user is loading the diags. Therefore, the group and policies change while we are restoring the diags.

Solution
========
Make sure that all db operations for restoring diags happen in one transaction. This happens by defining a new class that extends DiagnosticsHandlerImportable class and starts a transaction when restoring process begins and sending context for transaction to all restore calls for different diagnostics restore processes.