[OM-61960] Allow zero resize capacity for Commodity whose percentile is zero.
Diagnosis

In Market component, when a Sold Commodity's resize demand is 0, the resize quantity is set to the capacity and hence preventing a size down. This has led to a case at the customer Allergan where for a large but underutilized VM, no Scale down action is being generated.

Summary of Fix

In order to fix, there needs to be a distinction between 2 cases:
1. Reported utilization data is 0 - causing the percentile to be 0.
2. Utilization data and used value are not reported by probe.

To make this distinction, the following changes have been made to retain the fact that probe has not gathered metrics for a Commodity:
1. When Mediation SDK CommodityDTO does not have used value set, XL's CommoditySoldDTO also does not have used value set.
2. In Percentile dataset collection, add data point only if real-time used value is set for a Commodity.
3. In Percentile calculation (PercentileEditor) - do not set HistUtilizationValue field if the used AND utilizationData fields are unset.

Review: https://rbcommons.com/s/VMTurbo/r/41293/