[OM-59892]: Fix an issue that results tp diags load result in NPE
Intent
=====
Fix an NPE issue.

Problem
=======
When loading TP diags, diags load fail with NPE

Diagnosis
=========
In change with RB 41256 . I change the diags interface so it can accept a context. If there is no context, the context is send as null. The context is mostly not used. One specific case is PersistentMatchingAttributesIdentityStore as used in different components. For group component, it was a context while in topology processor, it does not. Therefore, null is send to method which result in NPE.

Solution
========
Use the local dsl context if the context sent is null. This way when an context is sent to restore diags, we will use it and if no context was sent, we will use local context like before.