[OM-61229]: RB-41330 Use action translation rather than original action for finding associated policies
Intent
======
Fix the issue that changing the policy type from RTA to SNOW does not revoke the accepted actions by RTA policy

Problem
=======
If the user goes and edits an policy with some accepted actions to an SNOW with RTA action. It is expected that action acceptance by RTA policy gets revoked and actions sent to ServiceNow.

Diagnosis
=========
To explain the solution you need to know the following background. The actions coming from the analysis component are kept in the recommendation field of an action object in action orchestration. These actions are then translated and the translation is kept in action translation field of the action. For each action, we keep a list of association policies which their change result in clearing associated actions. The problem was is the logic which finds the associated policies, it used the action before translation rather than action after translation. For VCPU resize, the action before translation is based on MHz and action after translation is based on cores. Therefore, we used the value of MHz and compared it with max core which obviously was bigger. Therefore, we used the policy for resizes above max rather than policy for resizes between threshold. As that policy has not changed we do not remove the acceptance for the action.

Solution
========
Update the logic for finding associated policy to use translated action.