[OM-62198]: Do not try to accept action from action approval manager if it is already accepted
Intent
======
Removing the log message that spammed the log

Problem
========
When setting up SNOW with RTA policy. After accepting an action, there was some log message in the log indicating that only ready actions can be accepted while the action mode is accepted.

Diagnosis
=========
When acceptance come back from service now, we try to execute the action. In that logic, we check if the action is ready and if so and RTA was in  place we persist the acceptance. We also change the mode of action to accepted. The next time the acceptance come from service now, we want to do the same thing. However, this time the action is accepted and not ready. Therefore, we throw an exception saying that the action is is not in the ready state.

Solution
=========
When the actions coming back from SNOW, we check if the the action is already accepted and if so we don't try to accept it again.