[OM-58398]: RB-41427 Expose cpu model in API
Intent
======
Expose cpu model in API so customer does not need to go to VCenter or HyperV to find it.

Background
==========
Customers need to know their CPU Model when creating a PhysicalMachine template. They will also need to know the CPU Model when creating a VM template when OM58398 is finished.

Solution
========
Take the cpuModel from entity properties, filled in by the Hypervisor probes and place it in the PM's aspects.

Related Fixes
============
1. OM62346: PhysicalMachineAspectMapper.mapEntityToAspect NPE