[OM-61086] Converted Prometheus config manager script into a sidecar.
It used to be a cron job, which was a better fit for some action that
just needs to happen periodically, but existing instances don't have the
necessary permissions given to `t8c-operator` to create cron jobs.
(Interestingly enough, it can create jobs. Something that is rarely done
directly)

The reason I didn't catch this earlier was that I had run a script from
product trust during the initial setup of the instance to turn on
Prometheus and Grafana. It turns out although it was claimed the ability
to enable Prometheus was there, the necessary permissions were
not and the PT script was giving permission to do anything to operator
in an attempt to work around that.

- Removed cron job and persistent volume claim templates.
- Added a new sidecar container configuration to Prometheus server pod.
- Changed `prometheus-merged-config` volume from persistent volume claim
  to `emptyDir`. No persistence is needed since this file is recreated
  every minute from scratch. The old contents are used if they exist but
  aren't expected to be there.
- Made the period between runs an argument to the script and wrapped the
  contents into an infinite loop. If the script fails with an error,
  Kubernetes should be able to restart the container.
- Removed a comment line that was no longer true.

Currently the latest version of the Docker image packaging the script is
specified. That's because I haven't figured out how to get access to the
XL version specified in cr.yaml from this `values.yaml` file yet.

Built and deployed operator and prometheus-config-manager on a recent
instance setup by Guy and verified that:
- There are no errors in operator logs about creating a cronjob. Before
  this change, I was getting such errors on this instance.
  I did see some errors about using `charts.helm.k8s.io/v1` instead of
  `charts.helm.k8s.io/v1alpha1` in `xl-release`, but these seemed more
  likely to be caused by the version used in
  `build/kubernetes/operator/deploy/crds/charts_v1alpha1_xl_cr-64gb-openjdk.yaml`
  and
  `build/kubernetes/operator/deploy/crds/charts_v1alpha1_xl_cr-turbo-with-turbo.yaml`
  than by my changes. They went away once I applied `charts_v1alpha1_xl_cr.yaml`.
- The persistent volume claim for configuration manager isn't created
  any more.
- I could turn telemetry on and off from the UI twice and the DataCloud
  gateway would start and stop sending data as expected.
- When an unhandled exception terminates the script, Kubernetes restarts
  the container as expected.
- The `configuration-manager` container isn't restarted by Kubernetes,
  nor does it exit spontaneously when no errors occur, or when it
  encounters a situation it can mitigate (e.g. When group and/or auth
  components are down. Warnings were printed to the logs in these cases).
- The timestamps in the messages in the logs about the script finishing
  its work are about 60s apart as desired. Sometimes they were a few
  more seconds apart but exact timing isn't required.
- When enough arguments aren't supplied to the script it prints a
  message and terminates with an error code. Kubernetes restarts it and
  the message is the desired one.