[OM-61064] Reverted "Made Prometheus server enabled by default."
I was let on the information that turbo-on-turbo uses Prometheus and is
turned on by default on new instances. So all the changes aimed at
reducing the overhead of Prometheus have to be reverted.

In addition to the above, Endre doesn't want Prometheus to be enabled by
default in operator but just in OVAs because the satellite components of
Prometheus require privileges we might not have when deploying XL on
clusters we don't own. There seems to be ways for satellite components
to work with reduced privileges and/or for Prometheus server to be
deployed without the satellite components or any special privileges, but
until I can test this properly I'm leaving Prometheus off by default in
t8c-operator. It will still be enabled by default in our new OVAs. The
old ones are the problem.

This reverts commit 48039fe1527ec6c102759ba3ec11582506f30c83.