[OM-62399]: RB:41485 Fix container proveit UI charts
**Problem**
Container ProveIt UI charts were not displaying because of a failed API call. The API call was failing because we were requesting entity stats for a set of commodities without providing any commodities.

**Solution**
The commodities list was empty because it was being assembled by a list of reason commodities pulled off of compound actions on a merged action. However, it looks like the commit that added the risk data structures containing these reason commodities was not merged in with the Container ProveIt UI branch.

This fix re-adds the code to provide the risk data structures.

**Testing Done**
Observed that the ProveIt UI charts now load correctly.
Rebuilt and re-ran tests in affected XL components.