[OM-62539]: RB:41516 Generate same recommendation OID for atomic actions
Sort the target OIDs and the commodity types in the merged resizes to generate the
unique ActionInfoModel for each atomic resize. This will ensure that the OID for the
atomic actions is the same even if the same market resize actions are in different
order in the market plan in differnent recommendation cycles.