[OM-62194]: Fix race condition in history topology coordination
Review RB-41537: https://rbcommons.com/s/VMTurbo/r/41537/

A race condition in inter-thread signaling could cause history
component to fail to process a received topology for up to twenty
minutes until an independent time-out kicked in, causing the topology
to be discarded and allowing the component to get back on track. This
was sometimes causing projected plan topologies to experience long
execution delays.