[OM-62219]: XL Stats: Do not show stats for clusters that don't exist.
Issue
=====
The /stats API returns data for clusters without display names.

Background
==========
Cluster retrieval happens in two steps:
1) Get the latest stats for all clusters from the history component (depending on the input query).
2) Retrieve auxiliary information (e.g. display names) for clusters from the group component.
3) Join them together.

If some clusters returned by 1) are not found in 2), we still return them, without the auxiliary information set - for example, setting the display name to the ID. This looks bad in the API, and the returned cluster ID is meaningless.

Fix
===
Drop stats for clusters that are not found in the system. This shouldn't happen often. Unfortunately this could affect pagination in certain corner cases, but that's not a big issue because we don't have any heavy pagination use cases for cluster stats.

Also, when the input is a list of explicit scope IDs and one of them does not exist, we shouldn't fail the whole request.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/41484/