[OM-62577]Support AD User in Multiple Groups (rb41525)
We discover all groups from AD and match them to external groups in Turbo
If a user is in more than one group we grant them the least permission role they are associated with
We grant them a combined scope for all of the groups they are in

Acceptance criteria:

User is in Group A with Observer role and one cluster for scope and Group B with Deployer role and no scope, they are granted Observer role for the entire environment
User is in Group A with automater role and one cluster for scope and Group B with observer role for a different cluster, they are granted Observer role for both clusters
User is in Group A with advisor role and one VM group, Group B with observer role for one business application group and Group C with Automater role for two clusters, they are granted Observer role for the VM group, business application group and two clusters.

Design:

Introduce a new feature flag enableMultiADGroupSupport, it's enabled (disable by default) the above use cases will be supported.

Testing Done
Add integration and unit test.
Tntegration testing results:
=================
1. two role, one Deployer, one observer (with scope) result should be observer with no scope empty.
Result JWT token:
{"sub":"corp\gary.zeng","IpAddress":"10.10.10.1","ROLES":["OBSERVER"],"UUID":"3261075146016","PROVIDER":"LDAP"}

User is in Group A with automater role and one cluster for scope and Group B with observer role for a different cluster, they are granted Observer role for both clusters.
Result JWT token:
{"sub":"corp\gary.zeng","IpAddress":"10.10.10.1","ROLES":["OBSERVER"],"UUID":"326