[OM-62530]: XL: Fix scoped observer action lists.
Issue
=====
Users scoped to groups were seeing actions and savings out of scope.

Background
==========
When testing whether a live action is accessible by a user we check to see if there are entities involved in the action that are not "static cloud entities" (e.g. TIERs) and are in the entities accessible by the user.

The problem was that when calculating "static cloud entities" in the user's scope we were iterating over all entities (by type) in the users scope and "union"-ing the static cloud types together. However, the implementation of union in Array and RoaringBitmap oid sets had a bug when a set was being "unioned" with a null - it was returning empty, where it should have been returning the set itself (treating null as empty).

Fix
===
- The fix is pretty straightforward - return "this" instead of "EMPTY_SET".
- I also added the option to check access without throwing an exception. For scoped users with lots of actions it could be an issue to continuously throw and catch exceptions.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/41598/