[OM-61086] Moved config-manager specification to the template section.
Moved part of the specification of the Prometheus configuration manager
sidecar of Prometheus server pod from the `values.yaml` of Prometheus
Helm chart to the corresponding `server.yaml` Helm template file.

Although we are generally avoiding to change the templates of upstream
Helm charts, this change was warranted by the fact that we were unable
to find a way to tie the version of the sidecar with that of the rest of
XL by changes to just `values.yaml`. It seems that we require a Helm
feature similar to https://github.com/helm/helm/issues/2133 or
https://github.com/helm/helm/pull/6876.

Built and deployed updated operator image on an instance in the lab.
Although, `7.22.6` was used as the tag in `xl/values.yaml` in operator
source, the `7.22.9-SNAPSHOT` tag from `cr.yaml` on the instance was
used for the sidecar in the Prometheus server deployment as verified by
using `kubectl describe` on the deployment and pod. That was the desired
behaviour.