[OM-61086] Updated Python script to account for changes in internal APIs.
Roman made breaking changes to internal API of auth component that can
be found at https://rbcommons.com/s/VMTurbo/r/41228. These broke the
Python script that attempted to get the customer e-mail from the license.
This change updates the script to work with the new DTO format.

Built and deployed updated version of prometheus-config-manager image to
a lab instance. Tested that it works with just one Turbonomic license
installed. Then installed a Grafana license from Roman and tested that
it continues to work. During testing I was looking at the logs and
merged Prometheus configuration to assert correct behaviour.