[OM-62536] Restored the PVC originally created for the cron job.
Although Prometheus configuration manager is now a sidecar of Prometheus
server pod, the easiest (and probably safest and most efficient) way to
create a persistent and unique `instance_id` is to create it where it's
going to be used. i.e. in the Prometheus configuration manager container.
But for it to be truly persistent `emptyDir` volumes aren't enough: we
need a persistent volume.

There are some changes compared to the original volume. Most notably,
access mode is now `ReadWriteOnce`.

The Hecketi minimum volume size of 1GB is still an issue, but not for
long, as there are plans to remove GlusterFS from our XL instances soon.

Built and deployed updated operator image to an instance in our lab and
watched the persistent volume claim and persistent volume to be created.
Looked at the logs of containers in Prometheus server pod and the
contents of the file that contains the merged Prometheus configuration.
Everything looks in order.