[OM-62536] Modified Prometheus config manager to manage `instance_id`.
Now prometheus configuration is persisted on a Kubernetes persistent
volume and a random unique (with very high probability) `instance_id` in
the UUID format is created when XL starts for the first time and
persisted in this volume.

Since `prometheus.yml` has to include this UUID and no other part of XL
currently needs to access it, it is only saved in this file.

Built and deployed updated versions of t8c-operator and
prometheus-config-manager images on 3 instances in our lab and verified
that:

- At first deployment the persistent volume and claim are created and
  `instance_id` is generated.
- All 3 instances had different `instance_id`s.
- The data sent to DataCloud from the 2 instances included the
  `instance_id`s for those instances. (I didn't wait for the 3rd)
- Letting prometheus-config-manager script to run multiple iterations
  didn't change the value of `instance_id`.
- Deleting the Prometheus server pod (and letting it be recreated
  automatically) didn't change the `instance_id`.
- Creating an XL instance with the new operator as well as disabling
  Prometheus, updating t8c-operator and re-enabling Prometheus work, but
  just updating t8c-operator and letting it reconcile Prometheus doesn't.
  It used to work too until recently. I believe I'm hit by OM-61676 or
  https://github.com/operator-framework/operator-sdk/issues/3329, but in
  order to verify this I need an instance where `updateAnnotations.sh`
  hasn't already run and the instances I had in my disposal didn't
  qualify.