[OM-62616]: XL Reports: Migrate existing "administrator" user to have the "Admin" role.
Issue
=====
Instances upgraded to 7.22.9 don't see the "Reports" button in Grafana.

Background
==========
In Grafana only organizational admins have access to scheduled reports. We currently create an "Editor" user for the Turbo "administrator" user, so there is no way for a customer to see the Grafana "Reports" button.

Fix
===
When initializing Grafana at extractor startup, ensure that the "administrator" user - if it exists - has the "Admin" role in the default organization. This is done via a GET, followed by a PATCH, on the organizational API. This is temporary code, and will be deleted once everyone is migrated to 7.22.9.

Also, we wanted to extend the timeout for report panel rendering to 5 minutes to allow long-running queries to finish. We also wanted this timeout to be overrideable via the config map. The easiest way to do this is to copy a "defaults.ini" file into the Grafana container and set our shipped-with default to what we want. Grafana supports two config files - the default, and a custom one which overrides the default. The custom one maps to the congfigmap injected by the operator, so we can't modify it within the image.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/41621/