[OM-62772] Disable Cloud Volume Scale actions by default for private preview.
Requirements

In order to have a private preview for the Cloud Storage Actions feature, the Cloud Storage Actions must be disabled by default. As a part of this Story, change the default value for the enableScaleActions setting to false.

Changes

Changed the default value of the enableScaleActions setting to false.

Review: https://rbcommons.com/s/VMTurbo/r/41719/