[OM-62824]: RB:41733 Disable resize actions on Container commodities with inconsistent capacities in a consistent scaling group
**Problem**
At an increasing number of customers we are seeing nonsense actions (ie an action to resize up VCPU when percentile utilization is at 2%) because members of a consistent scaling group have inconsistent capacities on the same commodity.

**Background**
Consistent scaling groups are defined as groups of entities that are required to scale together (ie they must have the same capacity on their commodities). Usually these are replicas defined by some shared configuration (ie container replicas defined by a shared ContainerSpec, members of the same VM autoscaling group that are required to share the same template, etc). When consistent scaling group members violate this assumption, the market attempts to scale them all to the same (highest) value of any of the group members.

Unfortunately in the case of containers, the shared configuration for CPU commodities is specified in millicore units (1/1000th of a core) whereas the market performs sizing actions in MHz. This means that when containers are running on nodes with different CPU speeds, even though they have the same configuration in millicores in kubernetes, they have different MHz speeds in the market, resulting in actions that make no sense.

**Solution**
While ultimately we should come up with a way to handle this properly, the right way to do so is very unclear. So for the moment, we will disable resize actions when we detect the scenario described above to prevent nonsense actions at the customer. No actions in this situation are better than nonsense actions that make no sense and can't be correctly executed anyway.

This is the short term solution The long term solution should be part of our epic of work to correctly handle millicores in Turbo.

Also add a helpful stage summary for the EphemeralEntityEditor pipeline stage to describe the edits this stage makes to the topology.

**Testing Done**
* Added new unit tests to cover new functionality.
* Verified the new stage summary in the logs and confirmed on a customer topology with the problem that we are correctly disabling resize actions on problematic groups.

EphemeralEntityHistoryStage --- WARNING (completed in 0.051s)
    Adjusted 6235 commodities.
    Disabled resize on 1228 commodities because of insufficient data.
    Disabled resize on 330 consistent scaling groups (of 1883 total) due to inconsistent capacities.