[OM-62564]: API/AO: Optimizing Action Stats Retrieval.
Intent
======
Speed up the POST /action/stats call used by Top Business Applications (and other) widgets.

Background
==========
The top accounts widget was taking several seconds at BofA scale. The longest API call was POST /action/stats group by action type. Upon closer investigation, we end up doing quite a lot of RPC calls back and forth between the API, repository, and action orchestrator. The intent of this review is to reduce the number of RPC calls in the average case - particularly for multiple users.

Implementation
==============
Action Orchestrator
===================
- The InvolvedEntitiesExpander currently makes RPC calls to the repository to determine when entities for input queries are "under" an ARM entity. We can use the Action Orchestrator's local topology graph to dramatically speed up those queries and eliminate the need for RPC calls.

API
===
In (**EntityScopeFactory** had an input of a set of oid tuples and an output of an expanded set of oids for each tuple in the input. The algorithm in both places was the same:
1) Iterate over the tuples.
2) For each tuple, make some kind of RPC call.

The change is to do "bulk" lookups - or, when possible, use the ApiId utility to remove the need for a lookup altogether if the information has already been found and cached in the API component.

Note - this required a change to allow "bulk" expansion in the **SupplyChainFetcherFactory** as well.

I also tried to be more consistent with using ApiId, and the information contained therein. We were already using it - and paying the cost for keeping information cached. To help with performance, you can now bulk-resolve **ApiIds** via the **UuidMapper**, which will minimize the number of calls required to 'resolve' multiple ApiIds to their entitiy/group/plan /target representations.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/41577/